First numerically functional example
Defined a simple model CifarNet8, trained it on CIFAR10, quantized it with PTQ for the digital core and
tested its accuracy. Less than 1% drop in accuracy. Included example
code for training and PTQ with plain pytorch (no pytorch lightning).
Made a few changes to make it work:

* Made BaseModules.py more user friendly
* Fixed bug introduced in previous commit where DIANAIdentity was also
  inserted AFTER each linear layer. Now this is only after the last
  layer.
* Prefer ceil rather than round for bitwidths >= 7 when determining
  power-of-two scales. Less accuracy drop for PTQ.
* Make qhparams init strategy more easy to configure since minmax is better
  for PTQ that meanstd (which is best for QAT).
* Set dataset scale to 1.0. Modifying this should be avoided in the
  future. Better remove the requantizer at the input and eps tunnel at
  the output in the future to work with int8 inputs/outputs.

NOTE: residuals and analog core not tested yet